Repositoryのinterfaceと実装を分離
