+++ repository/gorm/camp.go
@@ -1,0 +1,60 @@
+package gorm
+
+import (
+	"errors"
+
+	"github.com/go-sql-driver/mysql"
+	"gorm.io/gorm"
+
+	"github.com/traP-jp/rucQ/backend/model"
+)
+
+func (r *Repository) CreateCamp(camp *model.Camp) error {
+	if err := r.db.Create(camp).Error; err != nil {
+		var mysqlErr *mysql.MySQLError
+
+		if errors.As(err, &mysqlErr) && mysqlErr.Number == 1062 {
+			return model.ErrAlreadyExists
+		}
+
+		return err
+	}
+
+	return nil
+}
+
+func (r *Repository) GetCamps() ([]model.Camp, error) {
+	var camps []model.Camp
+
+	if err := r.db.Find(&camps).Error; err != nil {
+		return nil, err
+	}
+
+	return camps, nil
+}
+
+func (r *Repository) GetCampByID(id uint) (*model.Camp, error) {
+	var camp model.Camp
+
+	if err := r.db.Where(&model.Camp{
+		Model: gorm.Model{
+			ID: id,
+		},
+	}).First(&camp).Error; err != nil {
+		return nil, err
+	}
+
+	return &camp, nil
+}
+
+func (r *Repository) UpdateCamp(campID uint, camp *model.Camp) error {
+	if err := r.db.Where(&model.Camp{
+		Model: gorm.Model{
+			ID: campID,
+		},
+	}).Updates(camp).Error; err != nil {
+		return err
+	}
+
+	return nil
+}
