+++ repository/gorm/repository.go
@@ -1,0 +1,11 @@
+package gorm
+
+import "gorm.io/gorm"
+
+type Repository struct {
+	db *gorm.DB
+}
+
+func NewGormRepository(db *gorm.DB) *Repository {
+	return &Repository{db: db}
+}
