+++ repository/repository.go
@@ -1,11 +1,12 @@
 package repository
 
-import "gorm.io/gorm"
-
-type Repository struct {
-	db *gorm.DB
-}
-
-func NewRepository(db *gorm.DB) *Repository {
-	return &Repository{db: db}
+type Repository interface {
+	AnswerRepository
+	CampRepository
+	EventRepository
+	OptionRepository
+	QuestionRepository
+	QuestionGroupRepository
+	RoomRepository
+	UserRepository
 }
