+++ router/answers.go
@@ -3,78 +3,11 @@ package router
 import (
 	"net/http"
 
-	"github.com/jinzhu/copier"
 	"github.com/labstack/echo/v4"
-
-	"github.com/traP-jp/rucQ/backend/repository"
 )
 
-// GetMyAnswer アンケート回答取得
-// (GET /users/me/answers)
-func (s *Server) GetMyAnswer(e echo.Context, questionID QuestionId, params GetMyAnswersParams) error {
-	if params.XForwardedUser == nil {
-		return echo.NewHTTPError(http.StatusBadRequest, "X-Forwarded-User header is required")
-	}
-	user, err := s.repo.GetOrCreateUser(*params.XForwardedUser)
-
-	if err != nil {
-		e.Logger().Errorf("failed to get or create user: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	answer, err := s.repo.GetOrCreateAnswer(&repository.GetAnswerQuery{
-		QuestionID: uint(questionID),
-		UserID:     user.ID,
-	})
-
-	if err != nil {
-		e.Logger().Errorf("failed to get or create answer: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	_, err = s.repo.GetQuestionByID(uint(questionID))
-
-	if err != nil {
-		e.Logger().Errorf("failed to get question: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	var res AnswerBody
-
-	if err := copier.Copy(&res, &answer); err != nil {
-		e.Logger().Errorf("failed to copy model to response: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	// if answer.Content != nil {
-	// 	// TODO: model.QuestionTypeMultiple と model.QuestionTypeSingle の実装をapi.gen.goの型定義に合わせて修正する
-	// 	if question.Type == string(QuestionTypeMultiple) {
-	// 	 if err := res.Content.FromAnswerContent1(*answer.Content); err != nil {
-	// 	  e.Logger().Errorf("failed to convert content: %v", err)
-	//
-	// 	  return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	// 	 }
-	// 	} else {
-	// 	 if err := res.Content.FromAnswerContent0((*answer.Content)[0]); err != nil {
-	// 	  e.Logger().Errorf("failed to convert content: %v", err)
-	//
-	// 	  return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	// 	 }
-	// 	}
-	// }
-
-	// res.UserTraqId = *params.XForwardedUser // TODO: AnswerBody に UserTraqId がないためコメントアウト
-
-	return e.JSON(http.StatusOK, res)
-}
-
 // PutAnswer アンケート回答編集
-// (PUT /users/me/answers)
-func (s *Server) PutAnswer(e echo.Context, questionID QuestionId, params PutAnswerParams) error {
+func (s *Server) PutAnswer(e echo.Context, answerID AnswerId, params PutAnswerParams) error {
 	if params.XForwardedUser == nil {
 		return echo.NewHTTPError(http.StatusBadRequest, "X-Forwarded-User header is required")
 	}
@@ -84,32 +17,32 @@ func (s *Server) PutAnswer(e echo.Context, questionID QuestionId, params PutAnsw
 		return e.JSON(http.StatusBadRequest, err)
 	}
 
-	user, err := s.repo.GetOrCreateUser(*params.XForwardedUser)
+	// user, err := s.repo.GetOrCreateUser(*params.XForwardedUser)
 
-	if err != nil {
-		e.Logger().Errorf("failed to get or create user: %v", err)
+	// if err != nil {
+	// 	e.Logger().Errorf("failed to get or create user: %v", err)
 
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
+	// 	return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
+	// }
 
-	answer, err := s.repo.GetOrCreateAnswer(&repository.GetAnswerQuery{
-		QuestionID: uint(questionID),
-		UserID:     user.ID,
-	})
+	// answer, err := s.repo.GetOrCreateAnswer(&repository.GetAnswerQuery{
+	// 	QuestionID: uint(answerID),
+	// 	UserID:     user.ID,
+	// })
 
-	if err != nil {
-		e.Logger().Errorf("failed to get or create answer: %v", err)
+	// if err != nil {
+	// 	e.Logger().Errorf("failed to get or create answer: %v", err)
 
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
+	// 	return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
+	// }
 
-	_, err = s.repo.GetQuestionByID(uint(questionID))
+	// _, err = s.repo.GetQuestionByID(uint(answerID))
 
-	if err != nil {
-		e.Logger().Errorf("failed to get question: %v", err)
+	// if err != nil {
+	// 	e.Logger().Errorf("failed to get question: %v", err)
 
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
+	// 	return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
+	// }
 
 	// TODO: リクエストボディの型が変わったため、以下の処理を修正する
 	// if req.Content != nil {
@@ -134,19 +67,19 @@ func (s *Server) PutAnswer(e echo.Context, questionID QuestionId, params PutAnsw
 	//  }
 	// }
 
-	if err := s.repo.UpdateAnswer(answer); err != nil {
-		e.Logger().Errorf("failed to update answer: %v", err)
+	// if err := s.repo.UpdateAnswer(answer); err != nil {
+	// 	e.Logger().Errorf("failed to update answer: %v", err)
 
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
+	// 	return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
+	// }
 
 	var res AnswerBody // TODO: Answer を AnswerBody に修正
 
-	if err := copier.Copy(&res, &answer); err != nil {
-		e.Logger().Errorf("failed to copy model to response: %v", err)
+	// if err := copier.Copy(&res, &answer); err != nil {
+	// 	e.Logger().Errorf("failed to copy model to response: %v", err)
 
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
+	// 	return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
+	// }
 
 	// TODO: リクエストボディの型が変わったため、以下の処理を修正する
 	// if answer.Content != nil {
@@ -165,193 +98,6 @@ func (s *Server) PutAnswer(e echo.Context, questionID QuestionId, params PutAnsw
 	// }
 
 	// res.UserTraqId = *params.XForwardedUser // TODO: AnswerBody に UserTraqId がないためコメントアウト
-
-	return e.JSON(http.StatusOK, res)
-}
-
-// GetUserAnswer ユーザーのアンケート回答取得
-// (GET /users/{traQID}/answers)
-func (s *Server) GetUserAnswer(e echo.Context, traqID string, questionID QuestionId, params AdminGetAnswersParams) error { // TODO: GetUserAnswerParams を AdminGetAnswersParams に修正
-	if params.XForwardedUser == nil {
-		return echo.NewHTTPError(http.StatusBadRequest, "X-Forwarded-User header is required")
-	}
-	operator, err := s.repo.GetOrCreateUser(*params.XForwardedUser)
-
-	if err != nil {
-		e.Logger().Errorf("failed to get or create user: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	if !operator.IsStaff {
-		return echo.NewHTTPError(http.StatusForbidden, "Forbidden")
-	}
-
-	user, err := s.repo.GetOrCreateUser(traqID)
-
-	if err != nil {
-		e.Logger().Errorf("failed to get user: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	answer, err := s.repo.GetOrCreateAnswer(&repository.GetAnswerQuery{
-		QuestionID: uint(questionID),
-		UserID:     user.ID,
-	})
-
-	if err != nil {
-		e.Logger().Errorf("failed to get answer: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	var res AnswerBody // TODO: Answer を AnswerBody に修正
-
-	if err := copier.Copy(&res, &answer); err != nil {
-		e.Logger().Errorf("failed to copy model to response: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	_, err = s.repo.GetQuestionByID(uint(questionID))
-
-	if err != nil {
-		e.Logger().Errorf("failed to get question: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	// TODO: リクエストボディの型が変わったため、以下の処理を修正する
-	// if answer.Content != nil {
-	//  if question.Type == string(QuestionTypeMultiple) {
-	//   if err := res.Content.FromAnswerContent1(*answer.Content); err != nil {
-	//    e.Logger().Errorf("failed to convert content: %v", err)
-	//
-	//    return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	//   }
-	//  } else {
-	//   if err := res.Content.FromAnswerContent0((*answer.Content)[0]); err != nil {
-	//    e.Logger().Errorf("failed to convert content: %v", err)
-	//
-	//    return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	//   }
-	//  }
-	// }
-
-	// res.UserTraqId = string(traqID) // TODO: AnswerBody に UserTraqId がないためコメントアウト
-
-	return e.JSON(http.StatusOK, res)
-}
-
-// PutUserAnswer ユーザーのアンケート回答編集
-// (PUT /users/{traQID}/answers)
-func (s *Server) PutUserAnswer(e echo.Context, traqID string, questionID QuestionId, params AdminPutAnswerParams) error { // TODO: PutUserAnswerParams を AdminPutAnswerParams に修正
-	if params.XForwardedUser == nil {
-		return echo.NewHTTPError(http.StatusBadRequest, "X-Forwarded-User header is required")
-	}
-	operator, err := s.repo.GetOrCreateUser(*params.XForwardedUser)
-
-	if err != nil {
-		e.Logger().Errorf("failed to get or create user: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	if !operator.IsStaff {
-		return echo.NewHTTPError(http.StatusForbidden, "Forbidden")
-	}
-
-	var req AdminPutAnswerJSONRequestBody // TODO: PutUserAnswerJSONRequestBody を AdminPutAnswerJSONRequestBody に修正
-
-	if err := e.Bind(&req); err != nil {
-		return e.JSON(http.StatusBadRequest, err)
-	}
-
-	user, err := s.repo.GetOrCreateUser(traqID)
-
-	if err != nil {
-		e.Logger().Errorf("failed to get or create user: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	answer, err := s.repo.GetOrCreateAnswer(&repository.GetAnswerQuery{
-		QuestionID: uint(questionID),
-		UserID:     user.ID,
-	})
-
-	if err != nil {
-		e.Logger().Errorf("failed to get or create answer: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	_, err = s.repo.GetQuestionByID(uint(questionID))
-
-	if err != nil {
-		e.Logger().Errorf("failed to get question: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	// TODO: リクエストボディの型が変わったため、以下の処理を修正する
-	// if req.Content != nil {
-	//  if question.Type == string(QuestionTypeMultiple) {
-	//   content, err := req.Content.AsPutAnswerRequestContent1()
-	//
-	//   if err != nil {
-	//    return e.JSON(http.StatusBadRequest, err)
-	//   }
-	//
-	//   contentStrSlice := []string(content)
-	//   answer.Content = &contentStrSlice
-	//  } else {
-	//   content, err := req.Content.AsPutAnswerRequestContent0()
-	//
-	//   if err != nil {
-	//    return e.JSON(http.StatusBadRequest, err)
-	//   }
-	//
-	//   contentStrSlice := []string{string(content)}
-	//   answer.Content = &contentStrSlice
-	//  }
-	// }
-
-	if err := s.repo.UpdateAnswer(answer); err != nil {
-		e.Logger().Errorf("failed to update answer: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	var res AnswerBody // TODO: Answer を AnswerBody に修正
-
-	if err := copier.Copy(&res, &answer); err != nil {
-		e.Logger().Errorf("failed to copy model to response: %v", err)
-
-		return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	}
-
-	// TODO: リクエストボディの型が変わったため、以下の処理を修正する
-	// if answer.Content != nil {
-	//  if question.Type == string(QuestionTypeMultiple) {
-	//   if err := res.Content.FromAnswerContent1(*answer.Content); err != nil {
-	//    e.Logger().Errorf("failed to convert content: %v", err)
-	//
-	//    return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	//   }
-	//  } else {
-	//   if err := res.Content.FromAnswerContent0((*answer.Content)[0]); err != nil {
-	//    e.Logger().Errorf("failed to convert content: %v", err)
-	//
-	//    return echo.NewHTTPError(http.StatusInternalServerError, "Internal server error")
-	//   }
-	//  }
-	//
-	//  // res.UserTraqId = string(traqID) // TODO: AnswerBody に UserTraqId がないためコメントアウト
-	//
-	//  return e.JSON(http.StatusOK, res)
-	// }
 
 	return e.JSON(http.StatusOK, res)
 }
